reorganize, delete unused code
